+++ CHANGELOG.md
@@ -7,3 +7,6 @@
 ### Changed
 
 - Add Makefile and CHANGELOG.md
+- Fix bug of EventStream filter.
+- Rename `EventStream::into_json` to `json`
+- Change `EventStream::once` name and behavior.